+++ gems/quilt_rails/lib/quilt_rails.rb
@@ -2,7 +2,8 @@
 module Quilt
 end
 
-require "quilt_rails/engine"
 require "quilt_rails/version"
+require "quilt_rails/engine"
+require "quilt_rails/logger"
 require "quilt_rails/configuration"
 require "quilt_rails/react_renderable"