+++ gems/quilt_rails/lib/quilt_rails/logger.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Quilt
+  module Logger
+    def self.log(string)
+      if defined? Rails && Rails.logger.nil?
+        puts string
+      else
+        Rails.logger.info(string)
+      end
+    end
+  end
+end