+++ gems/quilt_rails/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.4.1] - 2019-08-22
+
+### Fixed
+
+- No longer breaks when used with `ShopifySecurityBase`
+
 ## [1.4.0] - 2019-08-21
 
 ### Added